requirements.txt
aiogram==3.13.1
aiohttp==3.10.10
python-dotenv==1.0.1
